tests/wpt/web-platform-tests/webstorage/storage_session_window_open.html

<!DOCTYPE HTML>
<html>
<head>
<title>WebStorage Test: sessionStorage - open a existing browsing context</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>

    test(function() {
        assert_true("sessionStorage" in window, "sessionStorage exist");

        var storage = window.sessionStorage;

        storage.setItem("FOO", "BAR");
        assert_equals(storage.getItem("FOO"), "BAR", "storage.getItem('FOO')");
        window.open("storage_session_window_open_second.html");

    }, "A new window to make sure there is a copy of the previous window's sessionStorage, and that they diverge after a change");

</script>
</body>
</html>
